+++ R/st_nn_pnt_proj.R
@@ -1,23 +1,19 @@
-.st_nn_pnt_proj = function(x, y, k, maxdist, progress) {
+.st_nn_pnt_proj = function(x, y, k, maxdist) {
 
   x_coord = sf::st_coordinates(x)
   y_coord = sf::st_coordinates(y)
 
-  # Progress bar
-  if(progress) pb = utils::txtProgressBar(min = 0, max = 1, initial = 0, style = 3)
-
   if(maxdist == Inf) {
-    nn = RANN::nn2(
+    nn = nabor::knn(
       query = x_coord,
       data = y_coord,
       k = k
       )
   } else {
-    nn = RANN::nn2(
+    nn = nabor::knn(
       query = x_coord,
       data = y_coord,
       k = k,
-      searchtype = "radius",
       radius = maxdist
     )
   }
@@ -37,7 +33,7 @@
   crs_units = st_crs(x)$units
 
   # Convert to meters
-  if(!is.na(crs_units)) {
+  if(!is.na(crs_units) & crs_units != "m") {
     dists = lapply(dists, units::set_units, crs_units, mode = "standard")
     dists = lapply(dists, units::set_units, "m", mode = "standard")
     dists = lapply(dists, as.numeric)
@@ -45,10 +41,6 @@
 
   # Remove names
   names(dists) = NULL
-
-  # Progress
-  if(progress) utils::setTxtProgressBar(pb, 1)
-  if(progress) cat("\n")
 
   # Return sparse lists
   return(list(ids, dists))